Backend/requirements.txt
Flask
googletrans==4.0.0-rc1
gTTS
SpeechRecognition
pyaudio
indic-nlp-library
flask-cors==5.0.0
python-dotenv==1.0.1